impelement autorization using gates

+++ app/Http/Controllers/QuestionsController.php
@@ -68,6 +68,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access Denied");
+            
+        }
         return view("questions.edit", compact('question'));
     }
 
@@ -93,6 +97,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access Denied");
+            
+        }
         $question->delete();
 
         return redirect('/questions')->with('success', 'Your question has been deleted');
